获取产品类别

+++ src/main/java/com/o2o/service/ShopCategoryService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ShopCategoryService {
 
-    public List<ShopCategory> queryShopCategoryByParentId (long parentId);
+    public List<ShopCategory> queryShopCategoryByParentId (int parentId);
 
 }
